refactor(routes): improve filenames ordering
